fix/script-DDL: update data types from MEDIUMINT to INT for address and phone tables since it doesn't exist in SSMS

+++ scripts/ddl/tabelas/script-DDL-tables.sql
@@ -1,9 +1,9 @@
 IF OBJECT_ID('address', 'U') IS NULL
 CREATE TABLE [address] (
-  [id_address] MEDIUMINT PRIMARY KEY IDENTITY(1,1) CHECK ([id_address] > 0),
+  [id_address] INT PRIMARY KEY IDENTITY(1,1) CHECK ([id_address] > 0),
   [street] VARCHAR(200) NOT NULL,
   [location] VARCHAR(200) NOT NULL,
-  [number] MEDIUMINT CHECK ([number] >= 0),
+  [number] INT CHECK ([number] >= 0),
   [reference] VARCHAR(200),
   [city] VARCHAR(100) NOT NULL,
   [state] VARCHAR(100) NOT NULL,
@@ -12,7 +12,7 @@ CREATE TABLE [address] (
 
 IF OBJECT_ID('phone', 'U') IS NULL
 CREATE TABLE [phone] (
-  [id_phone] MEDIUMINT PRIMARY KEY IDENTITY(1,1) CHECK ([id_phone] > 0),
+  [id_phone] INT PRIMARY KEY IDENTITY(1,1) CHECK ([id_phone] > 0),
   [nr_phone] VARCHAR(15) NOT NULL
 );
 
@@ -33,8 +33,8 @@ CREATE TABLE [users] (
   [dt_creation] DATETIME NOT NULL,
   [dt_birth] DATE NOT NULL,
   [specialty] VARCHAR(50),
-  [id_phone] MEDIUMINT NOT NULL CHECK ([id_phone] > 0),
-  [id_address] MEDIUMINT NOT NULL CHECK ([id_address] > 0),
+  [id_phone] INT NOT NULL CHECK ([id_phone] > 0),
+  [id_address] INT NOT NULL CHECK ([id_address] > 0),
   [id_user_type] TINYINT NOT NULL CHECK ([id_user_type] > 0),
   CONSTRAINT FK_Users_Phone FOREIGN KEY ([id_phone]) REFERENCES [phone] ([id_phone]),
   CONSTRAINT FK_Users_Address FOREIGN KEY ([id_address]) REFERENCES [address] ([id_address]),
